requirements.txt
Click==7.0
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
Pillow==7.0.0
Werkzeug==1.0.0
